BaseInstallerController: do not let 'crypto_passphrase' config values leak
When copying a configuration file to an installed system, we may also
expose a secret passphrase for an encrypted disk partition.

To fix this, remove all 'crypto_passphrase' values from configuration
data and directly dump this data to a configuration file on the
installed system instead of copying an original configuration file as
is.